COC-178-refactor(#111): study mock data 분리
