+++ src/mocks/data/study/createStudyData.ts
@@ -1,0 +1,20 @@
+export const createPublicStudyResponse = {
+  code: 1200,
+  message: '공개 스터디 생성에 성공했습니다.',
+  result: {
+    studyId: 1,
+  },
+};
+
+export const createPrivateStudyResponse = {
+  code: 1200,
+  message: '비공개 스터디 생성에 성공했습니다.',
+  result: {
+    studyId: 1,
+  },
+};
+
+export const createStudyErrorResponse = {
+  code: 4200,
+  message: '유효성 검사 실패',
+};
