+++ src/mocks/data/study/editStudyData.ts
@@ -1,0 +1,12 @@
+export const editStudyResponse = {
+  code: 1200,
+  message: '스터디 수정에 성공했습니다.',
+  result: {
+    studyId: 1,
+  },
+};
+
+export const editStudyErrorResponse = {
+  code: 4200,
+  message: '유효성 검사 실패',
+};
