+++ src/mocks/data/study/joinStudyData.ts
@@ -1,0 +1,19 @@
+export const joinPublicStudyResponse = {
+  code: 1200,
+  message: '공개 스터디 참가에 성공했습니다.',
+};
+
+export const joinPublicStudyErrorResponse = {
+  code: 4200,
+  message: '공개 스터디 참가에 실패했습니다.',
+};
+
+export const joinPrivateStudyResponse = {
+  code: 1200,
+  message: '비공개 스터디 참가에 성공했습니다.',
+};
+
+export const joinPrivateStudyErrorResponse = {
+  code: 4200,
+  message: '비공개 스터디 참가에 실패했습니다.',
+};
